requirements.txt
pyautogui>=0.9.54
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pynput>=1.7.6
pytesseract>=0.3.10

# DOM-based detection (optional, for Discord DevTools integration)
playwright>=1.40.0
